backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-marshmallow==0.15.0
flask-cors==4.0.0
marshmallow==3.21.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
minio==7.2.0
requests==2.31.0
flasgger==0.9.7.1
openai==1.12.0
qdrant-client>=1.7.0
numpy<2.0
httpx==0.26.0
beautifulsoup4==4.12.2
lxml==5.1.0
tiktoken==0.5.2
pandas==2.2.0
openpyxl==3.1.2
sqlalchemy==2.0.25

